move bundles to separate packages
