--- bundles/OdooWebsiteEditor/odoo-integration.ts
@@ -1,22 +1,0 @@
-import { BasicEditor } from '../BasicEditor/BasicEditor';
-import { DevTools } from '../../packages/plugin-devtools/src/DevTools';
-import { OdooWebsiteEditor } from './OdooWebsiteEditor';
-import { VRange, withRange } from '../../packages/core/src/VRange';
-import { DomLayoutEngine } from '../../packages/plugin-dom-layout/src/ui/DomLayoutEngine';
-import { Layout } from '../../packages/plugin-layout/src/Layout';
-import { Renderer } from '../../packages/plugin-renderer/src/Renderer';
-import { ImageNode } from '../../packages/plugin-image/src/ImageNode';
-import { getOdooCommands } from '../../packages/plugin-odoo/src/OdooCommands';
-
-export {
-    OdooWebsiteEditor,
-    BasicEditor,
-    DevTools,
-    Layout,
-    DomLayoutEngine,
-    Renderer,
-    ImageNode,
-    withRange,
-    VRange,
-    getOdooCommands,
-};
