+++ packages/bundle-basic-editor/BasicEditor.ts
@@ -1,0 +1,99 @@
+import JWEditor from '../core/src/JWEditor';
+import { Parser } from '../plugin-parser/src/Parser';
+import { Html } from '../plugin-html/src/Html';
+import { Char } from '../plugin-char/src/Char';
+import { LineBreak } from '../plugin-linebreak/src/LineBreak';
+import { Heading } from '../plugin-heading/src/Heading';
+import { Paragraph } from '../plugin-paragraph/src/Paragraph';
+import { List } from '../plugin-list/src/List';
+import { Indent } from '../plugin-indent/src/Indent';
+import { ParagraphNode } from '../plugin-paragraph/src/ParagraphNode';
+import { LineBreakNode } from '../plugin-linebreak/src/LineBreakNode';
+import { Span } from '../plugin-span/src/Span';
+import { Bold } from '../plugin-bold/src/Bold';
+import { Italic } from '../plugin-italic/src/Italic';
+import { Underline } from '../plugin-underline/src/Underline';
+import { Inline } from '../plugin-inline/src/Inline';
+import { Link } from '../plugin-link/src/Link';
+import { Divider } from '../plugin-divider/src/Divider';
+import { Image } from '../plugin-image/src/Image';
+import { Subscript } from '../plugin-subscript/src/Subscript';
+import { Superscript } from '../plugin-superscript/src/Superscript';
+import { Blockquote } from '../plugin-blockquote/src/Blockquote';
+import { Youtube } from '../plugin-youtube/src/Youtube';
+import { Table } from '../plugin-table/src/Table';
+import { Metadata } from '../plugin-metadata/src/Metadata';
+import { Renderer } from '../plugin-renderer/src/Renderer';
+import { Keymap } from '../plugin-keymap/src/Keymap';
+import { Align } from '../plugin-align/src/Align';
+import { Pre } from '../plugin-pre/src/Pre';
+import { TextColor } from '../plugin-textcolor/src/TextColor';
+import { BackgroundColor } from '../plugin-backgroundcolor/src/BackgroundColor';
+import { Layout } from '../plugin-layout/src/Layout';
+import { DomLayout } from '../plugin-dom-layout/src/DomLayout';
+import { DomEditable } from '../plugin-dom-editable/src/DomEditable';
+import { VNode } from '../core/src/VNodes/VNode';
+import { Input } from '../plugin-input/src/Input';
+import { Dialog } from '../plugin-dialog/src/Dialog';
+
+import template from './basicLayout.xml';
+import './basicLayout.css';
+
+export class BasicEditor extends JWEditor {
+    constructor() {
+        super();
+
+        this.configure({
+            defaults: {
+                Container: ParagraphNode,
+                Separator: LineBreakNode,
+            },
+            plugins: [
+                [Parser],
+                [Renderer],
+                [Layout],
+                [Keymap],
+                [Html],
+                [DomLayout],
+                [DomEditable],
+                [Inline],
+                [Char],
+                [LineBreak],
+                [Heading],
+                [Paragraph],
+                [List],
+                [Indent],
+                [Span],
+                [Bold],
+                [Italic],
+                [Underline],
+                [Link],
+                [Divider],
+                [Image],
+                [Subscript],
+                [Superscript],
+                [Blockquote],
+                [Youtube],
+                [Table],
+                [Metadata],
+                [Align],
+                [Pre],
+                [TextColor],
+                [BackgroundColor],
+                [Input],
+                [Dialog],
+            ],
+        });
+        this.configure(DomLayout, {
+            components: [
+                {
+                    id: 'editor',
+                    render(editor: JWEditor): Promise<VNode[]> {
+                        return editor.plugins.get(Parser).parse('text/html', template);
+                    },
+                },
+            ],
+            componentZones: [['editor', 'root']],
+        });
+    }
+}
