+++ packages/bundle-odoo-website-editor/OdooWebsiteEditor.ts
@@ -1,73 +1,73 @@
-import JWEditor from '../../packages/core/src/JWEditor';
-import { Parser } from '../../packages/plugin-parser/src/Parser';
-import { Html } from '../../packages/plugin-html/src/Html';
-import { Char } from '../../packages/plugin-char/src/Char';
-import { LineBreak } from '../../packages/plugin-linebreak/src/LineBreak';
-import { Heading } from '../../packages/plugin-heading/src/Heading';
-import { Paragraph } from '../../packages/plugin-paragraph/src/Paragraph';
-import { List } from '../../packages/plugin-list/src/List';
-import { Indent } from '../../packages/plugin-indent/src/Indent';
-import { ParagraphNode } from '../../packages/plugin-paragraph/src/ParagraphNode';
-import { LineBreakNode } from '../../packages/plugin-linebreak/src/LineBreakNode';
-import { Span } from '../../packages/plugin-span/src/Span';
-import { Bold } from '../../packages/plugin-bold/src/Bold';
-import { Italic } from '../../packages/plugin-italic/src/Italic';
-import { Underline } from '../../packages/plugin-underline/src/Underline';
-import { Inline } from '../../packages/plugin-inline/src/Inline';
-import { Link } from '../../packages/plugin-link/src/Link';
-import { Divider } from '../../packages/plugin-divider/src/Divider';
-import { Image } from '../../packages/plugin-image/src/Image';
-import { Subscript } from '../../packages/plugin-subscript/src/Subscript';
-import { Superscript } from '../../packages/plugin-superscript/src/Superscript';
-import { Blockquote } from '../../packages/plugin-blockquote/src/Blockquote';
-import { Youtube } from '../../packages/plugin-youtube/src/Youtube';
-import { Table } from '../../packages/plugin-table/src/Table';
-import { Metadata } from '../../packages/plugin-metadata/src/Metadata';
-import { Renderer } from '../../packages/plugin-renderer/src/Renderer';
-import { Keymap } from '../../packages/plugin-keymap/src/Keymap';
-import { Align } from '../../packages/plugin-align/src/Align';
-import { Pre } from '../../packages/plugin-pre/src/Pre';
-import { TextColor } from '../../packages/plugin-textcolor/src/TextColor';
-import { BackgroundColor } from '../../packages/plugin-backgroundcolor/src/BackgroundColor';
-import { Layout } from '../../packages/plugin-layout/src/Layout';
-import { DomLayout } from '../../packages/plugin-dom-layout/src/DomLayout';
-import { DomEditable } from '../../packages/plugin-dom-editable/src/DomEditable';
-import { VNode } from '../../packages/core/src/VNodes/VNode';
+import JWEditor from '../core/src/JWEditor';
+import { Parser } from '../plugin-parser/src/Parser';
+import { Html } from '../plugin-html/src/Html';
+import { Char } from '../plugin-char/src/Char';
+import { LineBreak } from '../plugin-linebreak/src/LineBreak';
+import { Heading } from '../plugin-heading/src/Heading';
+import { Paragraph } from '../plugin-paragraph/src/Paragraph';
+import { List } from '../plugin-list/src/List';
+import { Indent } from '../plugin-indent/src/Indent';
+import { ParagraphNode } from '../plugin-paragraph/src/ParagraphNode';
+import { LineBreakNode } from '../plugin-linebreak/src/LineBreakNode';
+import { Span } from '../plugin-span/src/Span';
+import { Bold } from '../plugin-bold/src/Bold';
+import { Italic } from '../plugin-italic/src/Italic';
+import { Underline } from '../plugin-underline/src/Underline';
+import { Inline } from '../plugin-inline/src/Inline';
+import { Link } from '../plugin-link/src/Link';
+import { Divider } from '../plugin-divider/src/Divider';
+import { Image } from '../plugin-image/src/Image';
+import { Subscript } from '../plugin-subscript/src/Subscript';
+import { Superscript } from '../plugin-superscript/src/Superscript';
+import { Blockquote } from '../plugin-blockquote/src/Blockquote';
+import { Youtube } from '../plugin-youtube/src/Youtube';
+import { Table } from '../plugin-table/src/Table';
+import { Metadata } from '../plugin-metadata/src/Metadata';
+import { Renderer } from '../plugin-renderer/src/Renderer';
+import { Keymap } from '../plugin-keymap/src/Keymap';
+import { Align } from '../plugin-align/src/Align';
+import { Pre } from '../plugin-pre/src/Pre';
+import { TextColor } from '../plugin-textcolor/src/TextColor';
+import { BackgroundColor } from '../plugin-backgroundcolor/src/BackgroundColor';
+import { Layout } from '../plugin-layout/src/Layout';
+import { DomLayout } from '../plugin-dom-layout/src/DomLayout';
+import { DomEditable } from '../plugin-dom-editable/src/DomEditable';
+import { VNode } from '../core/src/VNodes/VNode';
 
-import '../basicLayout.css';
-import { OdooSnippet } from '../../packages/plugin-odoo/src/OdooSnippet';
+import './odooLayout.css';
+import { OdooSnippet } from '../plugin-odoo/src/OdooSnippet';
 
-import { Toolbar } from '../../packages/plugin-toolbar/src/Toolbar';
-import { ParagraphButton } from '../../packages/plugin-heading/src/HeadingButtons';
-import { Heading1Button } from '../../packages/plugin-heading/src/HeadingButtons';
-import { Heading2Button } from '../../packages/plugin-heading/src/HeadingButtons';
-import { Heading3Button } from '../../packages/plugin-heading/src/HeadingButtons';
-import { Heading4Button } from '../../packages/plugin-heading/src/HeadingButtons';
-import { Heading5Button } from '../../packages/plugin-heading/src/HeadingButtons';
-import { Heading6Button } from '../../packages/plugin-heading/src/HeadingButtons';
-import { PreButton } from '../../packages/plugin-pre/src/PreButtons';
-import { BoldButton } from '../../packages/plugin-bold/src/BoldButtons';
-import { ItalicButton } from '../../packages/plugin-italic/src/ItalicButtons';
-import { UnderlineButton } from '../../packages/plugin-underline/src/UnderlineButtons';
-import { OrderedListButton } from '../../packages/plugin-list/src/ListButtons';
-import { UnorderedListButton } from '../../packages/plugin-list/src/ListButtons';
-import { IndentButton } from '../../packages/plugin-indent/src/IndentButtons';
-import { OutdentButton } from '../../packages/plugin-indent/src/IndentButtons';
-import { SaveButton } from '../../packages/plugin-odoo/src/SaveButton';
-import { HtmlNode } from '../../packages/plugin-html/src/HtmlNode';
-import { MediaButton } from '../../packages/plugin-odoo/src/MediaButton';
-import { CommandImplementation, CommandIdentifier } from '../../packages/core/src/Dispatcher';
-import { JWPlugin } from '../../packages/core/src/JWPlugin';
-import { OdooVideo } from '../../packages/plugin-odoo-video/src/OdooVideo';
-import { LinkButton } from '../../packages/plugin-odoo/src/LinkButton';
-import { DomZonePosition } from '../../packages/plugin-layout/src/LayoutEngine';
-import { HtmlDomRenderingEngine } from '../../packages/plugin-html/src/HtmlDomRenderingEngine';
+import { Toolbar } from '../plugin-toolbar/src/Toolbar';
+import { ParagraphButton } from '../plugin-heading/src/HeadingButtons';
+import { Heading1Button } from '../plugin-heading/src/HeadingButtons';
+import { Heading2Button } from '../plugin-heading/src/HeadingButtons';
+import { Heading3Button } from '../plugin-heading/src/HeadingButtons';
+import { Heading4Button } from '../plugin-heading/src/HeadingButtons';
+import { Heading5Button } from '../plugin-heading/src/HeadingButtons';
+import { Heading6Button } from '../plugin-heading/src/HeadingButtons';
+import { PreButton } from '../plugin-pre/src/PreButtons';
+import { BoldButton } from '../plugin-bold/src/BoldButtons';
+import { ItalicButton } from '../plugin-italic/src/ItalicButtons';
+import { UnderlineButton } from '../plugin-underline/src/UnderlineButtons';
+import { OrderedListButton } from '../plugin-list/src/ListButtons';
+import { UnorderedListButton } from '../plugin-list/src/ListButtons';
+import { IndentButton } from '../plugin-indent/src/IndentButtons';
+import { OutdentButton } from '../plugin-indent/src/IndentButtons';
+import { SaveButton } from '../plugin-odoo/src/SaveButton';
+import { HtmlNode } from '../plugin-html/src/HtmlNode';
+import { MediaButton } from '../plugin-odoo/src/MediaButton';
+import { CommandImplementation, CommandIdentifier } from '../core/src/Dispatcher';
+import { JWPlugin } from '../core/src/JWPlugin';
+import { OdooVideo } from '../plugin-odoo-video/src/OdooVideo';
+import { LinkButton } from '../plugin-odoo/src/LinkButton';
+import { DomZonePosition } from '../plugin-layout/src/LayoutEngine';
+import { HtmlDomRenderingEngine } from '../plugin-html/src/HtmlDomRenderingEngine';
 import {
     AlignLeftButton,
     AlignCenterButton,
     AlignRightButton,
     AlignJustifyButton,
-} from '../../packages/plugin-align/src/AlignButtons';
+} from '../plugin-align/src/AlignButtons';
 
 interface OdooWebsiteEditorOption {
     source: HTMLElement;
