+++ packages/bundle-odoo-website-editor/odooLayout.css
@@ -1,0 +1,4 @@
+jw-editor {
+    display: flex;
+    flex-direction: column;
+}
