+++ packages/plugin-devtools/test/devtools.test.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { BasicEditor } from '../../../bundles/BasicEditor/BasicEditor';
+import { BasicEditor } from '../../bundle-basic-editor/BasicEditor';
 import { VNode } from '../../core/src/VNodes/VNode';
 import { DevTools } from '../src/DevTools';
 import { click, nextTickFrame, keydown } from '../../utils/src/testUtils';
