+++ packages/plugin-fontawesome/test/FontAwesome.test.ts
@@ -1,6 +1,6 @@
 import { describePlugin } from '../../utils/src/testUtils';
 import { FontAwesome } from '../src/FontAwesome';
-import { BasicEditor } from '../../../bundles/BasicEditor/BasicEditor';
+import { BasicEditor } from '../../bundle-basic-editor/BasicEditor';
 import JWEditor from '../../core/src/JWEditor';
 import { Core } from '../../core/src/Core';
