+++ webpack-base.config.js
@@ -4,7 +4,10 @@ const glob = require('glob');
 const fs = require('fs');
 /* eslint-enable @typescript-eslint/no-var-requires */
 
-const integrationDevPath = path.resolve(__dirname, 'bundles/OdooWebsiteEditor/odoo-integration.ts');
+const integrationDevPath = path.resolve(
+    __dirname,
+    'packages/bundle-odoo-website-editor/odoo-integration.ts',
+);
 
 const mainConfig = {
     mode: 'development',
